requirements.txt
pyTelegramBotAPI
yfinance
pandas-ta
watchdog
flask
vnstock
paho-mqtt
python-dotenv
requests
psycopg2-binary
